adding recordlinker.utils.path for project_root and read_json
